fixed security issue

+++ backend/controllers/classroomController.js
@@ -14,37 +14,56 @@ const ClassController = {
   },
   deleteClass: async (req, res) => {
     try {
-      const { id } = req.body;
+      const { id, userSub } = req.body;
   
       // Use findByIdAndDelete to directly delete the class by its _id
-      const deletedClass = await Class.findByIdAndDelete(id);
-      await Post.deleteMany({classID: id})
-  
-      if (!deletedClass) {
-        return res.status(404).json({ message: 'Class not found' });
-      }
 
+      const doc = await Class.findById(id)
+
+      const docsUserSub = doc.userSub
+
+      if(docsUserSub === userSub ) {
+        const deletedClass = await Class.findByIdAndDelete(id);
+        await Post.deleteMany({classID: id})
+    
+        if (!deletedClass) {
+          return res.status(404).json({ message: 'Class not found' });
+        }
   
-      res.status(200).json({ message: 'Class deleted successfully' });
+    
+        res.status(200).json({ message: 'Class deleted successfully' });
+      } else {
+        res.status(403).json({message: "No auth"})
+      }
     } catch (error) {
       console.error('Error in deleting:', error);
       res.status(500).json({ message: `Server error: ${error}` });
     }
   },
   updateClass: async (req,res) => {
-    const {id, title, description, genre} = req.body
-    try {
-      const updatedClass = await Class.updateOne({_id: id}, {
-        $set : {
-          title,
-          description,
-          genre
-        }
-      })
-      res.status(200).json({message: updatedClass})
-    } catch (error) {
-      console.error('Error in updating', error)
-      res.status(500).json({ message: `Server error: ${error}` })
+    const {id, title, description, genre, userSub} = req.body
+
+    const doc = await Class.findById(id)
+
+    const docsUserSub = doc.userSub
+
+
+    if(docsUserSub === userSub) {
+      try {
+        const updatedClass = await Class.updateOne({_id: id}, {
+          $set : {
+            title,
+            description,
+            genre
+          }
+        })
+        res.status(200).json({message: updatedClass})
+      } catch (error) {
+        console.error('Error in updating', error)
+        res.status(500).json({ message: `Server error: ${error}` })
+      }
+    } else {
+      res.status(403).json({message: 'No auth'})
     }
   },
